+++ projects/packages/scheduled-updates/changelog/add-plugin-list-filter
@@ -1,0 +1,4 @@
+Significance: minor
+Type: added
+
+Scheduled Updates: Added a views link to filter plugins that are part of a scheduled update.
