+++ projects/packages/scheduled-updates/tests/php/class-scheduled-updates-test.php
@@ -557,19 +557,6 @@ class Scheduled_Updates_Test extends \WorDBless\BaseTestCase {
 	}
 
 	/**
-	 * Test get_scheduled_update_text.
-	 *
-	 * @dataProvider update_text_provider
-	 * @covers ::get_scheduled_update_text
-	 *
-	 * @param object $schedule The schedule object.
-	 * @param string $expected The expected text.
-	 */
-	public function test_get_scheduled_update_text( $schedule, $expected ) {
-		$this->assertSame( $expected, Scheduled_Updates::get_scheduled_update_text( $schedule ) );
-	}
-
-	/**
 	 * Test clear CRON cache.
 	 *
 	 * @covers ::clear_cron_cache
@@ -643,37 +630,6 @@ class Scheduled_Updates_Test extends \WorDBless\BaseTestCase {
 		$this->assertSame( 200, $result->get_status() );
 		$this->assertArrayNotHasKey( $id_1, $data );
 		$this->assertArrayHasKey( $id_2, $data );
-	}
-
-	/**
-	 * Data provider for test_get_scheduled_update_text.
-	 *
-	 * @return array[]
-	 */
-	public function update_text_provider() {
-		return array(
-			array(
-				(object) array(
-					'timestamp' => strtotime( 'next Monday 00:00' ),
-					'interval'  => WEEK_IN_SECONDS,
-				),
-				sprintf( 'Mondays at %s.', gmdate( get_option( 'time_format' ), strtotime( 'next Monday 8:00' ) ) ),
-			),
-			array(
-				(object) array(
-					'timestamp' => strtotime( 'next Tuesday 00:00' ),
-					'interval'  => DAY_IN_SECONDS,
-				),
-				sprintf( 'Daily at %s.', gmdate( get_option( 'time_format' ), strtotime( 'next Tuesday 8:00' ) ) ),
-			),
-			array(
-				(object) array(
-					'timestamp' => strtotime( 'next Sunday 00:00' ),
-					'interval'  => WEEK_IN_SECONDS,
-				),
-				sprintf( 'Sundays at %s.', gmdate( get_option( 'time_format' ), strtotime( 'next Sunday 8:00' ) ) ),
-			),
-		);
 	}
 
 	/**
